custar.pl: check if file being archived does now grow
if it happened in custar.pl execution, the resulting
archive would be broken. if one used that part in
some other code, it could be enlarging itself.

other change that does not affect custar.pl execution
is that if in tarlisted_open() one opened to stdout
(with '-' as file name), compression option is (also)
effective -- which it used not to be (one used to get
uncompressed ustar-formatted file in all cases).