A: tarxzin.pl
tarxzin.pl reads filenames to be archive from standard input,
filters (removes if any) and transforms (filenames) those.
then the (remaining, transformed) filenames are sorted, and
the files accessed using their original name are archived to
xz-compressed ustar-formatted file.

the main point of the existence of this that the sorting is
done after transformations.