A: cpaxgtar.sh
cpaxgtar.sh creates reproducible tar archives using
the Pax Interchange Format whenever ustar format is
not enough to handle the case. Such a cases are e.g.
filename (or linkname) too short to fit in ustar headers,
or one desires to use sparse feature to store empty
parts of files. This uses gtar(1) to do the actual
archiving (like custar.sh, sharing much of its code),
which may choose add other PaxHeaders if it figured out
a need for it.